requirements.txt
Django==2.2.5
djangorestframework==3.10.3
psycopg2==2.8.3
django-cors-headers==3.0.2
pytz==2019.2
sqlparse==0.3.0
gunicorn==19.9.0
whitenoise==3.3.1
django-heroku==0.3.1
dj-database-url==0.5.0
psycopg2-binary==2.8.3